refactor: move `send_commit()` to `PendingConversation` and `ConversationGuard`
Turns out it's better to offer two separate methods, because we're sending commits in
two distinct scenarios:
- External join commit, when we don't have a functioning conversation yet
- All other commits
