+++ crypto/src/mls/conversation/pending_conversation.rs
@@ -9,10 +9,10 @@ use crate::mls::conversation::conversation_guard::decrypt::buffer_messages::Mess
 use crate::mls::credential::crl::{extract_crl_uris_from_group, get_new_crl_distribution_points};
 use crate::mls::credential::ext::CredentialExt as _;
 use crate::prelude::{
-    ConversationId, MlsBufferedConversationDecryptMessage, MlsConversation, MlsConversationConfiguration,
-    MlsConversationDecryptMessage, MlsCustomConfiguration,
+    ConversationId, MlsBufferedConversationDecryptMessage, MlsCommitBundle, MlsConversation,
+    MlsConversationConfiguration, MlsConversationDecryptMessage, MlsCustomConfiguration,
 };
-use crate::{KeystoreError, LeafError, MlsError, RecursiveError};
+use crate::{KeystoreError, LeafError, MlsError, MlsTransportResponse, RecursiveError};
 use core_crypto_keystore::CryptoKeystoreMls as _;
 use core_crypto_keystore::entities::{MlsPendingMessage, PersistedMlsPendingGroup};
 use log::trace;
@@ -78,6 +78,28 @@ impl PendingConversation {
             .await
             .map_err(KeystoreError::wrap("saving mls pending groups"))
             .map_err(Into::into)
+    }
+
+    /// Send the commit via [crate::MlsTransport] and handle the response.
+    pub(crate) async fn send_commit(&self, commit: MlsCommitBundle) -> Result<()> {
+        let transport = self
+            .context
+            .mls_transport()
+            .await
+            .map_err(RecursiveError::root("getting mls transport"))?;
+        let transport = transport.as_ref().ok_or::<Error>(
+            RecursiveError::root("getting mls transport")(crate::Error::MlsTransportNotProvided).into(),
+        )?;
+
+        match transport
+            .send_commit_bundle(commit.clone())
+            .await
+            .map_err(RecursiveError::root("sending commit bundle"))?
+        {
+            MlsTransportResponse::Success => Ok(()),
+            MlsTransportResponse::Abort { reason } => Err(Error::MessageRejected { reason }),
+            MlsTransportResponse::Retry => Err(Error::CannotRetryWithoutConversation),
+        }
     }
 
     /// If the given message is the commit generated by [CentralContext::join_by_external_commit],
